+++ libs/aws/sabi-s3/src/actions/put_object/mod.rs
@@ -1,8 +1,8 @@
 mod error;
 pub use error::Error;
 
-mod file;
-pub use file::FileRequest;
+mod request;
+pub use request::FileRequest;
 
 mod response;
 pub use response::{Response, ResponseHeaders};
